% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bw_decomp.R
\name{bw_decomp}
\alias{bw_decomp}
\title{between-within decomposition of lifespan inequality measures}
\usage{
bw_decomp(
  age,
  ax,
  dx,
  lx,
  ex,
  distribution_type = "aad",
  prop = rep(1/ncol(dx), ncol(dx)),
  method = c("theil", "edag", "var", "mld", "gini", "mad", "aid", "H")
)
}
\arguments{
\item{age}{numeric vector of lower age bounds.}

\item{ax}{numeric matrix of the average time spent in the age interval of those dying within the interval with age in rows and subgroups in columns.}

\item{dx}{numeric matrix of the lifetable death distribution with age in rows and subgroups in columns.}

\item{lx}{numeric matrix of the lifetable survivorship with age in rows and subgroups in columns.}

\item{ex}{numeric matrix of remaining life expectancy with age in rows and subgroups in columns.}

\item{distribution_type}{\code{"aad"} for age-at death, \code{"rl"} for remaining life}

\item{prop}{numeric vector of starting fractions for each of the subgroups.}

\item{method}{character one of \code{"theil", "edag","var","mld","gini","mad","aid","H"}}
}
\description{
Partition a lifespan inequality index into additive components
of between-group inequality and within-group inequality. Presently
implemented for Theil's index, e-dagger, variance, mean log deviation,
the gini coeficient, mean absolute deviatation, the absolute inter-individual
difference, and life table entropy (H).
}
