% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{ineq_cov}
\alias{ineq_cov}
\title{ineq_cov}
\usage{
ineq_cov(age, dx, lx, ex, ax, distribution_type = c("aad", "rl"), check = TRUE)
}
\arguments{
\item{age}{numeric. vector of lower age bounds.}

\item{dx}{numeric. vector of the lifetable death distribution.}

\item{lx}{numeric. vector of the lifetable survivorship.}

\item{ex}{numeric. vector of remaining life expectancy.}

\item{ax}{numeric. vector of the average time spent in the age interval of those dying within the interval.}

\item{distribution_type}{character. Either \code{"aad"} (age at death) or \code{"rl"} (remaining life)}

\item{check}{logical. Shall we perform basic checks on input vectors? Default TRUE}
}
\description{
Calculate a lifetable column for the conditional coefficient of variation in lifetable ages at death
}
\details{
All input vectors must be the same length. Also, we recommend using input data from a life table by single year of age with a highest age group of at least age 110. If your data have a lower upper age bound, consider extrapolation methods, for instance a parametric Kannisto model (implemented in \code{MortalityLaws::MortalityLaw}). If your data are abridged, consider first smoothing over age, and calculating a life table by single year of age (for instance by smoothing with a pclm model in package \code{ungroup} or with a penalized B-spline approach in package \code{MortalitySmooth}).
}
\examples{

data(LT)
# A vector containing the conditional coefficient of variation in age at death 
CoV = ineq_cov(age=LT$Age,dx=LT$dx,lx=LT$lx,ex=LT$ex,ax=LT$ax)
# The coefficient of variation in age at death from birth
CoV[1]
# The coefficient of variation in age at death conditional upon survival to age 10
CoV[11]
}
\seealso{
\code{MortalityLaws::\link[MortalityLaws]{MortalityLaw}}

\code{ungroup::\link[ungroup]{pclm}}

\code{MortalitySmooth::\link[MortalitySmooth]{Mort1Dsmooth}}
}
