% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{ineq_drewnowski}
\alias{ineq_drewnowski}
\title{ineq_drewnowski}
\usage{
ineq_drewnowski(
  age,
  dx,
  ex,
  ax,
  distribution_type = c("aad", "rl"),
  check = TRUE
)
}
\arguments{
\item{age}{numeric. vector of lower age bounds.}

\item{dx}{numeric. vector of the lifetable death distribution.}

\item{ex}{numeric. vector of remaining life expectancy.}

\item{ax}{numeric. vector of the average time spent in the age interval of those dying within the interval.}

\item{distribution_type}{character. Either \code{"aad"} (age at death) or \code{"rl"} (remaining life)}

\item{check}{logical. Shall we perform basic checks on input vectors? Default TRUE}
}
\description{
This index is the simple complement of the gini coefficient, and we include it due to Aburto et al (2022)
}
\details{
All input vectors must be the same length. Also, we recommend using input data from a life table by single year of age with a highest age group of at least age 110. If your data have a lower upper age bound, consider extrapolation methods, for instance a parametric Kannisto model (implemented in \code{MortalityLaws::MortalityLaw}). If your data are abridged, consider first smoothing over age, and calculating a life table by single year of age (for instance by smoothing with a pclm model in package \code{ungroup} or with a penalized B-spline approach in package \code{MortalitySmooth}).

The formula for calculating the Gini was taken from the Shkolnikov (2010) spreadsheet, and is a simplification of the formulas described in Shkolnikov (2003) and Hanada (1983). This implementation allows the gini coefficient for both shortfall (remaining life) Shkolnikov (2010) and age-at-death (Permanyer 2019) distributions. This is the inverse of the Drewnowski index.
}
\examples{
data(LT)
# A vector containing the conditional age-at-death Drewnowski coefficients
D = ineq_drewnowski(age=LT$Age,dx=LT$dx,ex=LT$ex,ax=LT$ax, distribution_type = "aad")
#  A vector containing the conditional remaining life Gini coefficients
Dr = ineq_drewnowski(age=LT$Age,dx=LT$dx,ex=LT$ex,ax=LT$ax, distribution_type = "rl")
# To show how this relates to Gini:
G = ineq_gini(age=LT$Age,dx=LT$dx,ex=LT$ex,ax=LT$ax, distribution_type = "aad")
Gr = ineq_gini(age=LT$Age,dx=LT$dx,ex=LT$ex,ax=LT$ax, distribution_type = "rl")
\dontrun{
plot(0:110, Dr, type='l',col="red",ylab="conditional Gini",xlab="Age",ylim=c(0,1))
lines(0:110, D, col = "blue")
lines(0:110, Gr, col = "red", lty=2)
lines(0:110, G, col = "blue", lty=2)
legend("left",col = c("red","blue","red","blue"), lty=c(1,1,2,2),
       legend = c("remaining life Drewnowski","age at death Drewnowski",
                  "remaining life Gini", "age at death Gini"))
}
}
\references{
\insertRef{aburto2022drewnowski}{LifeIneq}
\insertRef{hanada1983}{LifeIneq}
\insertRef{shkolnikov2003}{LifeIneq}
\insertRef{shkolnikov2010}{LifeIneq}
}
