% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{ineq_eta_dag}
\alias{ineq_eta_dag}
\title{ineq_eta_dag}
\usage{
ineq_eta_dag(age, dx, lx, ex, ax, check = TRUE)
}
\arguments{
\item{age}{numeric. vector of lower age bounds.}

\item{dx}{numeric. vector of the lifetable death distribution.}

\item{lx}{numeric. vector of the lifetable survivorship.}

\item{ex}{numeric. vector of remaining life expectancy.}

\item{ax}{numeric. vector of the average time spent in the age interval of those dying within the interval.}

\item{check}{logical. Shall we perform basic checks on input vectors? Default TRUE}
}
\description{
Calculate a lifetable column for the average age at death lost at death of a population.
}
\details{
This quantity is not featured in the literature to our knowledge, but we've added it in order to make an age-at-death version of \eqn{e^\dagger} (which is a shortfall metric), for the sake of completeness. We also don't know what to call this measure, so we chose \code{eta_dag} to make the association with \code{edag} clear.
}
\examples{

data(LT)
# A vector containing the conditional mean age-at-death lost at death of a population
eaaddag = ineq_eta_dag(age=LT$Age,dx=LT$dx,lx=LT$lx,ex=LT$ex,ax=LT$ax)
# The aad-dag from birth
eaaddag[1]
# The aad-dag conditional upon survival to age 10
eaaddag[11]
\dontrun{
plot(0:110, eaaddag, type='l')
}
}
\seealso{
\code{MortalityLaws::\link[MortalityLaws]{MortalityLaw}}

\code{ungroup::\link[ungroup]{pclm}}

\code{MortalitySmooth::\link[MortalitySmooth]{Mort1Dsmooth}}
}
