% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{ineq_gini}
\alias{ineq_gini}
\title{ineq_gini}
\usage{
ineq_gini(age, lx, ex, ax)
}
\arguments{
\item{age}{numeric. vector of lower age bounds.}

\item{lx}{numeric. vector of the lifetable survivorship.}

\item{ex}{numeric. vector of remaining life expectancy.}

\item{ax}{numeric. vector of the average time spent in the age interval of those dying within the interval.}
}
\description{
Calculate a lifetable column for the conditional Gini coefficient of inequality in survivorship
}
\details{
All input vectors must be the same length. Also, we recommend using input data from a life table by single year of age with a highest age group of at least age 110. If your data have a lower upper age bound, consider extrapolation methods, for instance a parametric Kannisto model (implemented in \code{MortalityLaws::MortalityLaw}). If your data are abridged, consider first smoothing over age, and calculating a life table by single year of age (for instance by smoothing with a pclm model in package \code{ungroup} or with a penalized B-spline approach in package \code{MortalitySmooth}). 

  The formula for calculating the Gini was taken from the Shkolnikov (2010) spreadsheet, and is a simplification of the formulas described in Shkolnikov (2003) and Hanada (1983).
}
\examples{

data(LT)
# A vector containing the conditional Gini coefficients
G = ineq_gini(age=LT$Age,lx=LT$lx,ex=LT$ex,ax=LT$ax)
# The Gini coefficient from birth
G[1]
# The Gini coefficient conditional upon survival to age 10
G[11]
}
\references{
\insertRef{hanada1983}{LifeIneq}
\insertRef{shkolnikov2003}{LifeIneq}
\insertRef{shkolnikov2010}{LifeIneq}
}
\seealso{
\code{MortalityLaws::\link[MortalityLaws]{MortalityLaw}}

\code{ungroup::\link[ungroup]{pclm}}

\code{MortalitySmooth::\link[MortalitySmooth]{Mort1Dsmooth}}
}
