% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{ineq_iqr}
\alias{ineq_iqr}
\title{ineq_iqr}
\usage{
ineq_iqr(age, lx, n.grid = 1000)
}
\arguments{
\item{age}{numeric. vector of lower age bounds.}

\item{lx}{numeric. vector containing the lifetable survivorship. The starting radix (\eqn{\ell_{0}}) can be 100000, 1, or any value you like.}

\item{n.grid}{numeric. How many points do we want to interpolate? Defaults to 1000, which would work out to about a tenth of a year for a vector of life table survivors from age zero to 100+.}
}
\description{
Calculate the interquartile range survivorship age from a lifetable
}
\details{
The age and lx vectors must be the same length. This function estimates the interquartile range of the life table cohort by fitting a cubic spline through the survival curve. If your data have a lower upper age bound then that which corresponds to the top quartile of survivors, consider extrapolation methods, for instance a parametric Kannisto model (implemented in package 'MortalityLaws').
}
\examples{

data(LT)
# The iqr range of survival age
LTiqr <- ineq_iqr(age=LT$Age,lx=LT$lx,n.grid=1000)
LTiqr
}
