% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{ineq_MLD}
\alias{ineq_MLD}
\title{ineq_MLD}
\usage{
ineq_MLD(age, dx, lx, ex, ax)
}
\arguments{
\item{age}{numeric. vector of lower age bounds.}

\item{dx}{numeric. vector containing the lifetable deaths distribution.}

\item{lx}{numeric. vector containing the lifetable survivorship.}

\item{ex}{numeric. vector containing remaining life expectancy.}

\item{ax}{numeric. vector containing average time spent in age interval of those dying within the interval.}
}
\description{
Calculate a lifetable column for the conditional mean log deviation index of inequality in survivorship
}
\details{
All input vectors must be the same length. Also, we recommend using input data from a life table by single year of age with a highest age group of at least age 110. If your data have a lower upper age bound, consider extrapolation methods, for instance a parametric Kannisto model (implemented in package 'MortalityLaws'). If your data are abridged, consider first smoothing over age, and calculating a life table by single year of age (for instance by smoothing with a pclm model in package 'ungroup' or with a penalized B-spline approach in package 'MortalitySmooth').
}
\examples{

data(LT)
# A vector containing the conditional MLD indices
MLD = ineq_MLD(age=LT$Age,dx=LT$dx,lx=LT$lx,ex=LT$ex,ax=LT$ax)
# The MLD from birth
MLD[1]
# The MLD conditional upon survival to age 10
MLD[11]
}
