% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{ineq_quantile_lower}
\alias{ineq_quantile_lower}
\title{ineq_quantile_lower}
\usage{
ineq_quantile_lower(age, lx, quantile = 0.5)
}
\arguments{
\item{age}{numeric. vector of lower age bounds.}

\item{lx}{numeric. vector of the lifetable survivorship.}

\item{quantile}{numeric. value between zero and one indicating the quantile desired.}
}
\description{
Calculate quantiles of survivorship from a lifetable. Not vectorized: this function just calculates for the lowest age in the \code{lx} vector given.
}
\details{
Any of the quantile-based functions in
}
\examples{

data(LT)
# The median age at survival
LTmedian <- ineq_quantile_lower(age=LT$Age,lx=LT$lx,quantile=0.5)
LTmedian
# The age reached by 90\% of the life table cohort (i.e. the top 10\%)
ineq_quantile_lower(age=LT$Age,lx=LT$lx,quantile=0.1) 
# The difference between the bottom 10 and top 10 percent of survival age
ineq_quantile_lower(age=LT$Age,lx=LT$lx,quantile=0.1) -
ineq_quantile_lower(age=LT$Age,lx=LT$lx,quantile=0.9)
}
\seealso{
\code{MortalityLaws::\link[MortalityLaws]{MortalityLaw}}

\code{ungroup::\link[ungroup]{pclm}}

\code{MortalitySmooth::\link[MortalitySmooth]{Mort1Dsmooth}}
}
