% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{ineq_rel_eta_dag}
\alias{ineq_rel_eta_dag}
\title{ineq_rel_edag}
\usage{
ineq_rel_eta_dag(age, dx, lx, ex, ax, check = TRUE, distribution_type = "aad")
}
\arguments{
\item{age}{numeric. vector of lower age bounds.}

\item{dx}{numeric. vector of the lifetable death distribution.}

\item{lx}{numeric. vector of the lifetable survivorship.}

\item{ex}{numeric. vector of remaining life expectancy.}

\item{ax}{numeric. vector of the average time spent in the age interval of those dying within the interval.}

\item{check}{logical. Shall we perform basic checks on input vectors? Default TRUE}
}
\description{
Calculate a lifetable column for the elasticity of age at death, which is analogous to the Keyfitz-Leser \code{H} measure.
}
\details{
This method is implemented for the sake of completeness, since \eqn{e^\dagger} and \eqn{H} give the absolute and relative shortfall metrics, we've included \code{eta_dag} and \code{rel_eta_dag} to give age-at-death versions of these. We're not aware of anyone having used this formulation, and we do not offer a demographic interpretation of the scale of this metric, but we do point out that the conditional shape over age is qualitatively similar to other conditional relative measures from attainment (achieved age) distributions.
}
\examples{

data(LT)
# A vector containing the conditional rel_eta_dag values
H = ineq_rel_edag(age=LT$Age,dx=LT$dx,lx=LT$lx,ex=LT$ex,ax=LT$ax)
# The H from birth
H[1]
# The H conditional upon survival to age 10
H[11]
}
\references{
\insertRef{keyfitz1977mortality}{LifeIneq}
\insertRef{leser1955variations}{LifeIneq}
}
\seealso{
\code{MortalityLaws::\link[MortalityLaws]{MortalityLaw}}

\code{ungroup::\link[ungroup]{pclm}}

\code{MortalitySmooth::\link[MortalitySmooth]{Mort1Dsmooth}}
}
