% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LT}
\alias{LT}
\title{Year 2016 female Canadian single age lifetable from HMD}
\format{
A data.frame with 111 rows and 11 columns
}
\source{
\url{www.mortality.org}
}
\usage{
LT
}
\description{
Data taken from the file \code{fltper_1x1} using the default settings of the \code{HMDHFDplus} package.
}
\keyword{datasets}
