% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LTabr}
\alias{LTabr}
\title{Year 2016 female Canadian abridged age lifetable from HMD}
\format{
A data.frame with 24 rows and 11 columns
}
\source{
\url{https://mortality.org}
}
\usage{
LTabr
}
\description{
Data taken from the file \code{fltper_1x1} using the default settings of the \code{HMDHFDplus} package.
}
\details{
Data downloaded 9 Feb 2023.
}
\keyword{datasets}
