% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LTm}
\alias{LTm}
\title{Year 2016 male Canadian single age lifetable from HMD}
\format{
A data.frame with 111 rows and 11 columns
}
\source{
\url{https://mortality.org}
}
\usage{
LTm
}
\description{
Data taken from the file \code{mltper_1x1} using the default settings of the \code{HMDHFDplus} package.
}
\details{
Data downloaded 6 march 2024.
}
\keyword{datasets}
