% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{ineq}
\alias{ineq}
\title{calculate a lifespan inequality measure}
\usage{
ineq(
  age,
  dx,
  lx,
  ex,
  ax,
  method = c("var", "sd", "cov", "iqr", "aid", "gini", "mld", "edag", "cp", "Theil",
    "H", "mad"),
  ...
)
}
\arguments{
\item{age}{numeric. vector of lower age bounds.}

\item{dx}{numeric. vector of the lifetable death distribution.}

\item{lx}{numeric. vector of the lifetable survivorship.}

\item{ex}{numeric. vector of remaining life expectancy.}

\item{ax}{numeric. vector of the average time spent in the age}

\item{method}{one of \code{c("var","sd","cov","iqr","aid","gini","mld","edag","cp","theil","H","mad")}}

\item{...}{other optional arguments used by particular methods.}
}
\description{
Choose from variance \code{var}, standard deviation \code{sd}, coefficient of variation \code{cov}, interquartile range \code{iqr}, gini, absolute interindividual difference (absolute gini) \code{aid}, edagger, Kannisto's compression measure \code{cp}, Leser-Keyfitz entropy \code{H}, theil, mean log deviation \code{mld}, mean absolute deviation \code{mad} (wrt mean or median).
}
