% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{ineq}
\alias{ineq}
\title{calculate a lifespan inequality measure}
\usage{
ineq(
  age,
  dx,
  lx,
  ex,
  ax,
  method = c("variance", "sd", "iqr", "AID", "Gini", "MLD", "edag", "Cp"),
  ...
)
}
\arguments{
\item{age}{numeric. vector of lower age bounds.}

\item{dx}{numeric. vector of the lifetable death distribution.}

\item{lx}{numeric. vector of the lifetable survivorship.}

\item{ex}{numeric. vector of remaining life expectancy.}

\item{ax}{numeric. vector of the average time spent in the age}

\item{method}{one of \code{c("variance","sd","iqr","AID","Gini","MLD","edag","Cp")}}

\item{...}{other optional arguments used by particular methods.}
}
\description{
Choose from variance, standard deviation \code{sd},IQR, AID, Gini, edagger, or Kannisto's Cp.
}
