% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{ineq_AID}
\alias{ineq_AID}
\title{ineq_AID}
\usage{
ineq_AID(age, lx, ex, ax)
}
\arguments{
\item{age}{numeric. vector of lower age bounds.}

\item{lx}{numeric. vector containing the lifetable survivorship.}

\item{ex}{numeric. vector containing remaining life expectancy.}

\item{ax}{numeric. vector containing average time spent in age interval of those dying within the interval.}
}
\description{
Calculate a lifetable column for the conditional absolute inter-individual difference in lifespan (AID)
}
\details{
All input vectors must be the same length. Also, we recommend using input data from a life table by single year of age with a highest age group of at least age 110. If your data have a lower upper age bound, consider extrapolation methods, for instance a parametric Kannisto model (implemented in package 'MortalityLaws'). If your data are abridged, consider first smoothing over age, and calculating a life table by single year of age (for instance by smoothing with a pclm model in package 'ungroup' or with a penalized B-spline approach in package 'MortalitySmooth'). 

The formula for calculating the AID was taken from the citet{Shkolnikov2010} spreadsheet, and is a simplification of the formula described in citet{Shkolnikov2003}.
}
\examples{

data(LT)
# A vector containing the conditional absolute inter-individual difference in lifespan
AID = ineq_AID(age=LT$Age,lx=LT$lx,ex=LT$ex,ax=LT$ax)
# The absolute inter-individual difference in lifespan from birth
AID[1]
# The absolute inter-individual difference in lifespan from age 10
AID[11]
}
