% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{ineq_cp}
\alias{ineq_cp}
\title{ineq_cp}
\usage{
ineq_cp(age, lx, p = 0.5)
}
\arguments{
\item{age}{numeric. vector of lower age bounds.}

\item{lx}{numeric. vector of the lifetable survivorship.}

\item{p}{numeric. What proportion of the life table cohort do you want captured in the C measure? The default is .5}
}
\description{
Calculate Kannisto's C-measures from a lifetable
}
\details{
The age and lx vectors must be the same length. This function estimates the shortest distance between two ages containing p percent of the life table cohort's death. The mechanics behind the function are to fit a cubic spline through the survival curve to estimate surivorship between age intervals. If your data have an upper age bound lower than 110, consider extrapolation methods, for instance a parametric Kannisto model (implemented in package 'MortalityLaws').

This function is currently very slow because the author is a better demographer than she is programmer. For ideas of how to improve this function, please let email vanraalte@demogr.mpg.de.

The concept behind Kannisto's C-measures is found in Kannisto (2000)
}
\examples{

data(LT)
# The shortest age range containing half of the deaths
(C50 <- ineq_cp(age=LT$Age,lx=LT$lx,p=.5))
}
\references{
\insertRef{kannisto2000}{LifeIneq}
}
