% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{ineq_cp}
\alias{ineq_cp}
\title{ineq_cp}
\usage{
ineq_cp(age, lx, p = 0.5, check = TRUE)
}
\arguments{
\item{age}{numeric. vector of lower age bounds.}

\item{lx}{numeric. vector of the lifetable survivorship.}

\item{p}{numeric. What proportion of the life table cohort do you want captured in the C measure? The default is .5}

\item{check}{logical. Shall we perform basic checks on input vectors? Default TRUE}
}
\description{
Calculate Kannisto's C-measures from a lifetable
}
\details{
The \code{age} and \code{lx} vectors must be the same length. This function estimates the shortest distance between two ages containing \code{p} proportion of the life table cohort's death. The mechanics behind the function are to fit a monotonic cubic spline through the survival curve to estimate continuous surivorship between age intervals. If your data have an upper age bound lower than 110, consider extrapolation methods, for instance a parametric Kannisto model (implemented in package \code{MortalityLaws}).

The concept behind Kannisto's C-measures is found in Kannisto (2000)
}
\examples{

data(LT)
# The shortest age range containing half of the deaths
(C50 <- ineq_cp(age=LT$Age,lx=LT$lx,p=.5))
# The shortest age range containing 80\% the deaths
(C80 <- ineq_cp(age=LT$Age,lx=LT$lx,p=.8))
}
\references{
\insertRef{kannisto2000}{LifeIneq}
}
