% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{ineq_edag}
\alias{ineq_edag}
\title{ineq_edag}
\usage{
ineq_edag(age, dx, lx, ex, ax, check = TRUE)
}
\arguments{
\item{age}{numeric. vector of lower age bounds.}

\item{dx}{numeric. vector of the lifetable death distribution.}

\item{lx}{numeric. vector of the lifetable survivorship.}

\item{ex}{numeric. vector of remaining life expectancy.}

\item{ax}{numeric. vector of the average time spent in the age interval of those dying within the interval.}

\item{check}{logical. Shall we perform basic checks on input vectors? Default TRUE}
}
\description{
Calculate a lifetable column for the conditional life disparity (\eqn{e^\dagger}) of a population.
}
\details{
All input vectors must be the same length. Also, we recommend using input data from a life table by single year of age with a highest age group of at least age 110. If your data have a lower upper age bound, consider extrapolation methods, for instance a parametric Kannisto model (implemented in \code{MortalityLaws::MortalityLaw}). If your data are abridged, consider first smoothing over age, and calculating a life table by single year of age (for instance by smoothing with a pclm model in package \code{ungroup} or with a penalized B-spline approach in package \code{MortalitySmooth}).
}
\examples{

data(LT)
# A vector containing the conditional life disparity of a population
edag = ineq_edag(age=LT$Age,dx=LT$dx,lx=LT$lx,ex=LT$ex,ax=LT$ax)
# The life disparity from birth
edag[1]
# The life disparity conditional upon survival to age 10
edag[11]
\dontrun{
plot(0:110, edag, type='l')
}
}
\seealso{
\code{MortalityLaws::\link[MortalityLaws]{MortalityLaw}}

\code{ungroup::\link[ungroup]{pclm}}

\code{MortalitySmooth::\link[MortalitySmooth]{Mort1Dsmooth}}
}
