% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indices.R
\name{ineq_MAD}
\alias{ineq_MAD}
\title{ineq_variance}
\usage{
ineq_MAD(age, dx, lx, ex, ax, center_type = c("ex", "mean", "median"))
}
\arguments{
\item{age}{numeric. vector of lower age bounds.}

\item{dx}{numeric. vector of the lifetable death distribution.}

\item{lx}{numeric. vector of the lifetable survivorship.}

\item{ex}{numeric. vector of remaining life expectancy.}

\item{ax}{numeric. vector of the average time spent in the age interval of those dying within the interval.}

\item{center_type}{character either, \code{"ex"}, \code{"mean"} (same thing), or \code{"median"}}
}
\description{
Calculate a lifetable column for the conditional mean average deviation in lifetable ages at death. This may be with respect to either conditional life expectancy or conditional median remaining lifespan.
}
\details{
All input vectors must be the same length. Also, we recommend using input data from a life table by single year of age with a highest age group of at least age 110. If your data have a lower upper age bound, consider extrapolation methods, for instance a parametric Kannisto model (implemented in \code{MortalityLaws::MortalityLaw}). If your data are abridged, consider first smoothing over age, and calculating a life table by single year of age (for instance by smoothing with a pclm model in package \code{ungroup} or with a penalized B-spline approach in package \code{MortalitySmooth}).
}
\examples{

data(LT)
# A vector containing the conditional MAD in age at death 
M = ineq_MAD(age=LT$Age,dx=LT$dx,lx=LT$lx,ex=LT$ex,ax=LT$ax)
# The MAD in age at death from birth
M[1]
# The MAD in age at death conditional upon survival to age 10
M[11]
# MAD wrt condition median
ineq_MAD(age = LT$Age, dx = LT$dx,
  lx = LT$lx, ex = LT$ex, ax = LT$ax, 
  center_type = "median")
}
\seealso{
\code{MortalityLaws::\link[MortalityLaws]{MortalityLaw}}

\code{ungroup::\link[ungroup]{pclm}}

\code{MortalitySmooth::\link[MortalitySmooth]{Mort1Dsmooth}}
}
